cloud_storage: extract materializer from async manifest view
Split into smaller classes/responsibility.

No behavior change intended.
